# docs/swagger.yaml
basePath: /
definitions:
...
  pkg_http.JFrogPlatform:
    properties:
      architecture:
        type: string
      os:
        type: string
    type: object
  pkg_http.JFrogWebhook:
    properties:
      data:
        $ref: '#/definitions/pkg_http.JFrogWebhookData'
      domain:
        type: string
      event_type:
        type: string
      jpd_origin:
        type: string
      source:
        type: string
      subscription_key:
        type: string
    type: object
  pkg_http.JFrogWebhookData:
    properties:
      image_name:
        type: string
      name:
        type: string
      path:
        type: string
      platforms:
        items:
          $ref: '#/definitions/pkg_http.JFrogPlatform'
        type: array
      repo_key:
        type: string
      sha256:
        type: string
      size:
        type: integer
      tag:
        type: string
    type: object
...
  types.AuditLogStats:
    properties:
      approved:
        type: integer
      date:
        type: string
      rejected:
        type: integer
      updates:
        type: integer
      webhooks:
        type: integer
    type: object
...
  types.Repository:
    properties:
      digest:
        description: optional digest field
        type: string
      host:
        type: string
      name:
        type: string
      tag:
        type: string
    type: object
...
info:
  contact: {}
  description: HTTP API exposed by Keel. Admin routes are registered only when an
    authenticator is enabled. Provider webhooks require Basic or Bearer authorization
    only when authenticated webhooks are enabled; the registry webhook is always unauthenticated.
  title: Keel HTTP API
  version: "1.0"
paths:
...
  /v1/stats:
    get:
      description: Returns daily webhook, approval, rejection, and update counts.
        This route exists only when the authenticator is enabled.
      operationId: getStats
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.AuditLogStats'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Permission denied
          schema:
            type: string
        "500":
          description: Store query failed
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get audit statistics
      tags:
      - Admin
...
  /v1/webhooks/jfrog:
    post:
      consumes:
      - application/json
      description: Requires Basic or Bearer authorization only when authenticatedWebhooks
        is enabled.
      operationId: receiveJFrogWebhook
      parameters:
      - description: JFrog Docker push
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_http.JFrogWebhook'
      responses:
        "200":
          description: Accepted
        "400":
          description: Malformed payload, missing image name, or missing tag
          schema:
            type: string
        "401":
          description: Unauthorized when authenticated webhooks are enabled
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Receive a JFrog webhook
      tags:
      - Webhooks
  /v1/webhooks/native:
    post:
      consumes:
      - application/json
      description: Requires Basic or Bearer authorization only when authenticatedWebhooks
        is enabled.
      operationId: receiveNativeWebhook
      parameters:
      - description: Repository event
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.Repository'
      responses:
        "200":
          description: Accepted
        "400":
          description: Malformed repository, missing name, or missing tag
          schema:
            type: string
        "401":
          description: Unauthorized when authenticated webhooks are enabled
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Receive a native webhook
      tags:
      - Webhooks
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"